pyproject.toml
[project]
name = "django-hstore-field"
version = "0.0.1"
description = "An easy to use postgres hstore field that is based on django-hstore-widget"
readme = "README.md"
requires-python = ">=3.19"
dependencies = ["django-hstore-widget>=0.0.19"]

[dependency-groups]
build = ["hatch>=1.14.1"]
test = [
    "pytest>=8.0.1",
    "pytest-django>=4.8.0",
    "psycopg[binary]>=3",
    "selenium>=4.26.1",
]


[tool.hatch.build]
packages = ["src/django_hstore_field"]
directory = 'py-dist'


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
